+++ gh_repl.py
@@ -1,24 +1,18 @@
 #!/usr/bin/env python3
-"""Use this program to interact with your repos (note that there are better...
-
+"""Use this program to interact with your repos (note that there are better
 solutions out there)
 """
 from __future__ import annotations
 
 import json
-import os
 import sys
 from typing import Any, Callable
 
 from metprint import LogType
 
-import lib.githubREST as githubREST
-import lib.utils as utils
+from lib import github_rest, utils
 
-
-def clear():
-	"""Clear the terminal."""
-	os.system('cls' if os.name == 'nt' else 'clear')
+# pylint:disable=import-outside-toplevel
 
 
 def printMarkdown(raw: str, maxpages: int = 0):
@@ -26,7 +20,8 @@ def printMarkdown(raw: str, maxpages: int = 0):
 	try:
 		import pypandoc
 		from catpandoc import pandoc2ansi, processpandoc
-		output = json.loads(pypandoc.convert_text(raw, 'json', 'md'))
+
+		output = json.loads(pypandoc.convert_text(raw, "json", "md"))
 		pandoc = pandoc2ansi.Pandoc2Ansi(80, 0, (4, 0, 0))
 		markdown = ""
 		for block in output["blocks"]:
@@ -39,14 +34,16 @@ def printMarkdown(raw: str, maxpages: int = 0):
 
 def listRepos(data: str, user: str):
 	"""List the user repos."""
-	userRepos = githubREST.getListOfUserRepos(user, data)
-	paginatedList(userRepos, 8, githubREST.printRepo)
+	userRepos = github_rest.getListOfUserRepos(user, data)
+	paginatedList(userRepos, 8, github_rest.printRepo)
 
 
-def paginatedList(iterable: list[Any],
-perPage: int,
-printFunc: Callable[[dict[Any, Any]], None],
-maxpages: int = 0):
+def paginatedList(
+	iterable: list[Any],
+	perPage: int,
+	printFunc: Callable[[dict[Any, Any]], None],
+	maxpages: int = 0,
+):
 	"""Print a paginated list."""
 	totalPages = len(iterable) // perPage + 1
 	for index, iteration in enumerate(iterable):
@@ -54,7 +51,7 @@ maxpages: int = 0):
 		if maxpages != 0 and page >= maxpages:
 			return
 		if index > 0 and index % perPage == 0:
-			input("Page {} of {} (Next)>".format(index // perPage, totalPages))
+			input(f"Page {index // perPage} of {totalPages} (Next)>")
 		printFunc(iteration)
 
 
@@ -64,15 +61,12 @@ maxpages: int = 0):
 
 def replhelp():
 	"""Return help text for the REPL."""
-	clear()
-	utils.printf.logPrint("Most of the time the 'user' arg can be omitted",
-	LogType.INFO)
+	utils.clear()
+	utils.printf.logPrint("Most of the time the 'user' arg can be omitted", LogType.INFO)
 	utils.printf.logPrint("Functions: ", LogType.BOLD)
-	for function in functions:
-		utils.printf.logPrint("- {} : {}".format(
-		function,
-		list(functions[function].__code__.co_varnames[:functions[function].__code__
-		.co_argcount])))
+	for name, function in functions.items():
+		params = list(function.__code__.co_varnames[: function.__code__.co_argcount])
+		utils.printf.logPrint(f"- {name} : {params}")
 
 
 def replexit():
@@ -101,76 +95,76 @@ def watching(user: str | None = None):
 def profile(user: str | None = None):
 	"""Print user profile info."""
 	user = username if user is None else user
-	clear()
-	userData = githubREST.getUser(user)
-	utils.printf.logPrint("{}".format(userData["name"]), LogType.BOLD)
+	utils.clear()
+	userData = github_rest.getUser(user)
+	utils.printf.logPrint(f"{userData['name']}", LogType.BOLD)
 	utils.printf.logPrint(
-	"{}\nAvatar: {} \nCompany: {} \nLocation: {} \nEmail: {} \nFollowers: {} Following: {}"
-	.format(userData["login"],
-	userData["avatar_url"],
-	userData["company"],
-	userData["location"],
-	userData["email"],
-	userData["followers"],
-	userData["following"]))
+		f"""{userData['login']}
+Avatar: {userData['avatar_url']}
+Company: {userData['company']}
+Location: {userData['location']}
+Email: {userData['email']}
+Followers: {userData['followers']} Following: {userData['following']}"""
+	)
 
 
 def gists(user: str | None = None):
 	"""Print paginated list of user gists."""
 	user = username if user is None else user
-	userGists = githubREST.getUserGists(user)
-	paginatedList(userGists, 30, githubREST.printGist)
+	userGists = github_rest.getUserGists(user)
+	paginatedList(userGists, 30, github_rest.printGist)
 
 
 def showrepo(repo: str, user: str | None = None):
 	"""Print user repo data for a given repo."""
-	clear()
+	utils.clear()
 	user = username if user is None else user
-	rawMarkdown = githubREST.getReadme(user + "/" + repo)
-	repoText = githubREST.getRepo(user + "/" + repo)
-	githubREST.printRepo(repoText)
+	rawMarkdown = github_rest.getReadme(user + "/" + repo)
+	repoText = github_rest.getRepo(user + "/" + repo)
+	github_rest.printRepo(repoText)
 	utils.printf.logPrint("README", LogType.BOLD)
 	printMarkdown(rawMarkdown, 1)
 
 
 def showreadme(repo: str, user: str | None = None):
 	"""Print the readme for a given repo."""
-	clear()
+	utils.clear()
 	user = username if user is None else user
-	printMarkdown(githubREST.getReadme(user + "/" + repo))
+	printMarkdown(github_rest.getReadme(user + "/" + repo))
 
 
 def searchissues(searchTerm: str):
 	"""Search function for issues."""
-	issues = githubREST.search(searchTerm, context="issues")
-	paginatedList(issues, 30, githubREST.printIssue)
+	issues = github_rest.search(searchTerm, context="issues")
+	paginatedList(issues, 30, github_rest.printIssue)
 
 
 def searchrepos(searchTerm: str):
 	"""Search function for repos."""
-	searchRepos = githubREST.search(searchTerm, context="repositories")
-	paginatedList(searchRepos, 10, githubREST.printRepo)
+	searchRepos = github_rest.search(searchTerm, context="repositories")
+	paginatedList(searchRepos, 10, github_rest.printRepo)
 
 
 def searchusers(searchTerm: str):
 	"""Search function for users."""
-	users = githubREST.search(searchTerm, context="users")
-	paginatedList(users, 30, githubREST.printUser)
+	users = github_rest.search(searchTerm, context="users")
+	paginatedList(users, 30, github_rest.printUser)
 
 
 functions = {
-"exit": replexit,
-"help": replhelp,
-"repos": repos,
-"stars": stars,
-"watching": watching,
-"profile": profile,
-"showrepo": showrepo,
-"showreadme": showreadme,
-"searchissues": searchissues,
-"searchrepos": searchrepos,
-"searchusers": searchusers,
-"gists": gists}
+	"exit": replexit,
+	"help": replhelp,
+	"repos": repos,
+	"stars": stars,
+	"watching": watching,
+	"profile": profile,
+	"showrepo": showrepo,
+	"showreadme": showreadme,
+	"searchissues": searchissues,
+	"searchrepos": searchrepos,
+	"searchusers": searchusers,
+	"gists": gists,
+}
 
 
 def repl():